Extract enterprise feature flag module
